+++ src/Routes.tsx
@@ -5,29 +5,29 @@ import MenuLayout from "./layout/MenuLayout";
 import Login from "./pages/Login";
 import Home from "./pages/Home";
 import WishList from "./pages/WishList";
-import MyPage from "./pages/MyPage";
-import Category from "./pages/Category";
+import MyPage from "./pages/MyPage/MyPage";
+import Category from "./pages/Category/Category";
 import ProductDetail from "./pages/ProductDetail";
 import ReviewWritePage from "./pages/ReviewWritePage";
 import ScanPage from "./pages/ScanPage";
 import CartPage from "./pages/CartPage";
 import QRCodePage from "./pages/QRCodePage";
-import Signup from "./pages/Signup";
-import AboutYouPage from "./pages/AboutYouPage";
-import Survey from "./pages/SurveyPage";
-import Welcome from "./pages/Welcome";
+import Signup from "./pages/SignUp/Signup";
+import AboutYouPage from "./pages/SignUp/AboutYouPage";
+import Survey from "./pages/SignUp/SurveyPage";
+import Welcome from "./pages/SignUp/Welcome";
 import PurchaseList from "./pages/PurchaseListPage";
 import PurchaseDetail from "./pages/PurchaseDetailPage";
 import Withdraw from "./pages/WithdrawPage";
-import SettingLanguange from "./pages/MypageLanguange";
+import SettingLanguange from "./pages/MyPage/MypageLanguange";
 import IngredientDetail from "./pages/IngredientDetailPage";
 import ChatbotPage from "./pages/Chatbotpage";
 import MypagePersonalColor from "./pages/MypagePersonalColor";
 import MypageSkinType from "./pages/MypageSkinType";
 import MypageSkinConcern from "./pages/MypageSkinConcern";
-import MbtiTest from "./pages/MbtiTest";
-import MbtiQuestion from "./pages/MbtiQuestion";
-import CategoryList from "./pages/CategoryList";
+import MbtiTest from "./pages/SkinMBTI/MbtiTest";
+import MbtiQuestion from "./pages/SkinMBTI/MbtiQuestion";
+import CategoryList from "./pages/Category/CategoryList";
 
 const routes = [
   // DefaultLayout
