+++ src/components/Survey/SkipHeader.tsx
