+++ src/components/Survey/SurveyStep1.tsx
