+++ src/components/Survey/SurveyStep2.tsx
