+++ src/pages/SignUp/SurveyPage.tsx
@@ -1,13 +1,13 @@
 import styled from "styled-components";
-import Header from "../components/Header";
-import colors from "../styles/colors";
+import Header from "../../components/Header";
+import colors from "../../styles/colors";
 import { Link, useNavigate } from "react-router-dom";
 import { useState } from "react";
-import SurveyStep1 from "../components/SurveyStep1";
-import SurveyStep2 from "../components/SurveyStep2";
-import SurveyStep3 from "../components/SurveyStep3";
-import Button from "../components/Button";
-import StepIndicator from "../components/StepIndicator";
+import SurveyStep1 from "../../components/Survey/SurveyStep1";
+import SurveyStep2 from "../../components/Survey/SurveyStep2";
+import SurveyStep3 from "../../components/Survey/SurveyStep3";
+import Button from "../../components/Button";
+import StepIndicator from "../../components/StepIndicator";
 
 const Wrap = styled.div`
   width: 100vw;
