+++ src/pages/SignUp/Welcome.tsx
